Implement image_scripts.
On restart we look through /image_scripts and execute any files in that
directory. We only wait for them to start because some programs may
never exit. Output like other commands is sent to redis and eventually
to the connected user.

No longer write a custom .htaccess file to the service directory, since
this can now be down w/ image_scripts.

Process termination no longer relies on the command string to kill the
process and now does it properly via the language routines provided.

Also a bug is fixed that was prevalent in previous versions where if a
process start child processes those processes were never killed. Currently
this is specific to Linux and will fail to compile on other systems but
support can be easily added when cross platform support is needed.

Implement #25